+++ Jibberwock.Persistence.DataAccess/Commands/Products/CreateProduct.cs
@@ -28,7 +28,7 @@ namespace Jibberwock.Persistence.DataAccess.Commands.Products
         [Required]
         public Product Product { get; set; }
 
-        public CreateProduct(ILogger logger, User performedBy, string connectionId, int serviceId, string comment, Product product)
+        public CreateProduct(ILogger logger, User performedBy, string connectionId, long serviceId, string comment, Product product)
             : base(logger, performedBy, connectionId, serviceId, comment)
         {
             Product = product;
