+++ Jibberwock.Persistence.DataAccess/Commands/Security/DeleteSecurityGroup.cs
@@ -27,7 +27,7 @@ namespace Jibberwock.Persistence.DataAccess.Commands.Security
         [Required]
         public Group Group { get; set; }
 
-        public DeleteSecurityGroup(ILogger logger, User performedBy, string connectionId, int serviceId, string comment, Group group)
+        public DeleteSecurityGroup(ILogger logger, User performedBy, string connectionId, long serviceId, string comment, Group group)
             : base(logger, performedBy, connectionId, serviceId, comment)
         {
             Group = group;
